enemies/boss now use float numbers to move better, using x's and y's instead of bbox for all movement, include all headers in one header file for cleaner code

+++ src/graphics.h
@@ -26,4 +26,4 @@ typedef struct button_s
 SDL_Surface *load_Image (char *filename);
 
 /* Displaying Image function: Shows image on a given surface; makes image visible */
-void show_Surface (int x, int y, SDL_Surface* source, SDL_Surface* destination, SDL_Rect* clip);
+void show_Surface (float x, float y, SDL_Surface* source, SDL_Surface* destination, SDL_Rect* clip);
